+++ apps/modernjs/modern.config.ts
@@ -11,11 +11,24 @@ export default defineConfig({
   runtime: {
     router: true,
   },
+  security: {
+    checkSyntax: true,
+  },
+  source: {
+    // downgrade @module-federation related pkgs
+    include: [
+      // should set module-federation in outer repo
+      /universe\/packages/,
+    ],
+  },
   // source: {
   //   enableAsyncEntry: true,
   // },
   plugins: [appTools()],
   tools: {
+    babel(config) {
+      config.sourceType = 'unambiguous';
+    },
     webpack: (config, { webpack, appendPlugins }) => {
       if (config?.output) {
         config.output.publicPath = 'http://localhost:4001/';