+++ apps/runtime-demo/3006-runtime-remote/webpack.config.js
@@ -9,6 +9,7 @@ const path = require('path');
 const {
   ModuleFederationPlugin,
 } = require('@module-federation/enhanced/webpack');
+const packageJson = require('./package.json');
 
 module.exports = composePlugins(
   withNx(),
@@ -18,6 +19,11 @@ module.exports = composePlugins(
       ignored: ['**/node_modules/**', '**/@mf-types/**'],
     };
     // const ModuleFederationPlugin = webpack.container.ModuleFederationPlugin;
+    config.watchOptions = {
+      ignored: ['**/dist/**'],
+    };
+    config.devServer.host = '127.0.0.1';
+
     config.plugins.push(
       new ModuleFederationPlugin({
         name: 'runtime_remote1',
@@ -29,13 +35,32 @@ module.exports = composePlugins(
           './WebpackPng': './src/components/WebpackPng',
         },
         shared: {
-          lodash: {},
-          antd: {},
-          react: {},
-          'react/': {},
-          'react-dom': {},
-          'react-dom/': {},
+          lodash: {
+            singleton: true,
+            requiredVersion: '^4.0.0',
+          },
+          antd: {
+            singleton: true,
+            requiredVersion: '^4.0.0',
+          },
+          react: {
+            singleton: true,
+            requiredVersion: '^18.2.0',
+          },
+          'react/': {
+            singleton: true,
+            requiredVersion: '^18.2.0',
+          },
+          'react-dom': {
+            singleton: true,
+            requiredVersion: '^18.2.0',
+          },
+          'react-dom/': {
+            singleton: true,
+            requiredVersion: '^18.2.0',
+          },
         },
+        runtimePlugins: [path.join(__dirname, './runtimePlugin.ts')],
       }),
     );
     // config.externals={
@@ -57,6 +82,7 @@ module.exports = composePlugins(
     // e.g. `config.plugins.push(new MyPlugin())`
     config.output = {
       ...config.output,
+      publicPath: 'http://localhost:3006/',
       scriptType: 'text/javascript',
     };
     config.optimization = {