+++ packages/chrome-devtools/src/utils/sdk/graph.ts
@@ -31,7 +31,7 @@ export interface EdgeType {
   };
 }
 
-const validateSemver = (schema: string) => {
+export const validateSemver = (schema: string) => {
   // https://regex101.com/r/vkijKf/1
   const reg =
     /^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(\.(0|[1-9]\d*))?(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$/gm;
@@ -180,8 +180,9 @@ export class DependencyGraph {
     let info = name;
 
     const remote = this.snapshot[target];
-    if (remote && 'version' in remote) {
-      info += `:${remote.version}`;
+    if (remote && ('version' in remote || 'remoteEntry' in remote)) {
+      // @ts-expect-error
+      info += `:${remote.version || remote.remoteEntry}`;
     }
 
     if (!this.identifyMap.has(targetWithoutType)) {