+++ packages/dts-plugin/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @module-federation/dts-plugin
 
+## 0.1.8
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.8
+- @module-federation/managers@0.1.8
+- @module-federation/third-party-dts-extractor@0.1.8
+
+## 0.1.7
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.7
+- @module-federation/managers@0.1.7
+- @module-federation/third-party-dts-extractor@0.1.7
+
+## 0.1.6
+
+### Patch Changes
+
+- 72c7b80: chore: fix release tag
+- Updated dependencies [72c7b80]
+  - @module-federation/managers@0.1.6
+  - @module-federation/sdk@0.1.6
+  - @module-federation/third-party-dts-extractor@0.1.6
+
+## 0.1.5
+
+### Patch Changes
+
+- f9b8af7: feat: supports the use of @module-federation/enhanced/runtime and also has type hints
+- 71559fb: fix: output error in file
+- Updated dependencies [ca271ab]
+- Updated dependencies [1a9c6e7]
+  - @module-federation/sdk@0.1.5
+  - @module-federation/managers@0.1.5
+
+## 0.1.4
+
+### Patch Changes
+
+- 8f3a440: fix: detect whether the project is ts
+- Updated dependencies [8f3a440]
+- Updated dependencies [2f697b9]
+  - @module-federation/sdk@0.1.4
+  - @module-federation/managers@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes