+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,41 @@
 # [3.1.0](https://github.com/module-federation/universe/compare/utils-3.0.2...utils-3.1.0) (2023-10-26)
 
+## 3.0.13
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.8
+
+## 3.0.12
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.7
+
+## 3.0.11
+
+### Patch Changes
+
+- 72c7b80: chore: fix release tag
+- Updated dependencies [72c7b80]
+  - @module-federation/sdk@0.1.6
+
+## 3.0.10
+
+### Patch Changes
+
+- Updated dependencies [ca271ab]
+- Updated dependencies [1a9c6e7]
+  - @module-federation/sdk@0.1.5
+
+## 3.0.9
+
+### Patch Changes
+
+- Updated dependencies [8f3a440]
+- Updated dependencies [2f697b9]
+  - @module-federation/sdk@0.1.4
+
 ## 3.0.8
 
 ### Patch Changes